Add files via upload
This code is used to create a web application that allows the user to add items to a shopping list. The code is written in HTML, CSS and JavaScript. Used module to import Firebase libraries.

HTML code defines the structure and content of a web page, such as the title, image, input field, button, and list. The HTML code also links to external CSS and JavaScript files.
CSS code defines the style and appearance of a web page, such as font, color, layout and responsiveness. The CSS code is stored in a separate file called index.css.
JavaScript code defines the functionality and interactivity of the web page, such as adding click event listeners, getting input values, pushing items to the Firebase database, and adding items to a list. The JavaScript code is stored in a separate file called index.js and uses modules to import the Firebase libraries.